+++ forui/lib/src/theme/theme_data.dart
@@ -7,7 +7,7 @@ import 'package:forui/forui.dart';
 
 import 'package:forui/src/theme/breakpoints.dart';
 
-import 'package:forui/src/widgets/line_calendar.dart';
+import 'package:forui/src/widgets/line_calendar/line_calendar.dart';
 
 /// Defines the configuration of the overall visual [FTheme] for a widget subtree.
 ///
