+++ forui/lib/src/widgets/line_calendar/line_calendar.dart
@@ -9,54 +9,9 @@ import 'package:forui/forui.dart';
 import 'package:forui/src/foundation/tappable.dart';
 import 'package:forui/src/foundation/util.dart';
 
-//TODO: localizations.
-const _days = ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun'];
 final _start = (DateTime(1900), null);
 const _textSpacing = 2.0;
 
-bool _true(DateTime _) => true;
-
-typedef FlineSelection = (DateTime start, DateTime? end);
-
-class FlineController extends FCalendarController<DateTime?> {
-  final Predicate<DateTime> _selectable;
-  final FlineSelection? _period;
-
-  FlineController({
-    FlineSelection? period,
-    DateTime? initialSelection,
-    Predicate<DateTime>? selectable,
-  })  : assert(
-          period == null || (period.$1.isUtc && (period.$2?.isUtc ?? true)),
-          'value must be in UTC timezone',
-        ),
-        assert(
-          () {
-            if (period == null) {
-              return true;
-            }
-            final start = period.$1;
-            final end = period.$2;
-            if (end == null) {
-              return true;
-            }
-            return start.isBefore(end) || start.isAtSameMomentAs(end);
-          }(),
-          'end date must be greater than or equal to start date',
-        ),
-        _selectable = selectable ?? _true,
-        _period = period ?? _start,
-        super(initialSelection ?? DateTime.now());
-
-  @override
-  bool selectable(DateTime date) => _selectable(date);
-
-  @override
-  void select(DateTime date) => value = value?.toLocalDate() == date.toLocalDate() ? null : date;
-
-  @override
-  bool selected(DateTime date) => value?.toLocalDate() == date.toLocalDate();
-}
 
 /// A calendar that can be scrolled horizontally.
 class FLineCalendar extends StatefulWidget {
@@ -137,7 +92,7 @@ class _FLineCalendarState extends State<FLineCalendar> {
         itemExtent: size,
         itemBuilder: (context, index) => Container(
           padding: EdgeInsets.symmetric(horizontal: style.itemPadding),
-          child: _Tile(
+          child: FlineCalendarTile(
             style: style,
             selected: widget.selected,
             date: widget.start.add(Duration(days: index)),
@@ -155,96 +110,6 @@ class _FLineCalendarState extends State<FLineCalendar> {
   }
 }
 
-class _Tile extends StatelessWidget {
-  final FLineCalendarStyle style;
-  final ValueNotifier<LocalDate> selected;
-  final LocalDate date;
-  final bool today;
-
-  const _Tile({
-    required this.style,
-    required this.selected,
-    required this.date,
-    required LocalDate today,
-  }) : today = date == today;
-
-  Color _style(BuildContext context, bool selected) => switch ((selected, today)) {
-        (true, true) => style.selectedCurrentDateIndicatorColor,
-        (true, false) => const Color(0x00000000),
-        (false, true) => style.unselectedCurrentDateIndicatorColor,
-        (false, false) => const Color(0x00000000),
-      };
-
-  @override
-  Widget build(BuildContext context) => ValueListenableBuilder(
-        valueListenable: selected,
-        builder: (context, selectedDate, _) {
-          final isSelected = selectedDate == date;
-          return FTappable.animated(
-            semanticLabel: date.toString(),
-            selected: isSelected,
-            onPress: () => selected.value = date,
-            child: Stack(
-              children: [
-                Positioned.fill(
-                  child: DecoratedBox(
-                    decoration: isSelected ? style.selectedDecoration : style.unselectedDecoration,
-                    child: Column(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: [
-                        // TODO: replace with DefaultTextStyle.merge when textHeightBehavior has been added.
-                        merge(
-                          textHeightBehavior: const TextHeightBehavior(
-                            applyHeightToFirstAscent: false,
-                            applyHeightToLastDescent: false,
-                          ),
-                          style:
-                              isSelected ? style.content.selectedDateTextStyle : style.content.unselectedDateTextStyle,
-                          //TODO: Localization.
-                          child: Text('${date.day}'),
-                        ),
-                        const SizedBox(height: _textSpacing),
-                        // TODO: replace with DefaultTextStyle.merge when textHeightBehavior has been added.
-                        merge(
-                          textHeightBehavior: const TextHeightBehavior(
-                            applyHeightToFirstAscent: false,
-                            applyHeightToLastDescent: false,
-                          ),
-                          style: isSelected ? style.content.selectedDayTextStyle : style.content.unselectedDayTextStyle,
-                          child: Text(_days[date.weekday - 1]),
-                        ),
-                      ],
-                    ),
-                  ),
-                ),
-                Positioned(
-                  top: 6,
-                  right: 6,
-                  child: Container(
-                    height: 4,
-                    width: 4,
-                    decoration: BoxDecoration(
-                      color: _style(context, isSelected),
-                      shape: BoxShape.circle,
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          );
-        },
-      );
-
-  @override
-  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
-    super.debugFillProperties(properties);
-    properties
-      ..add(DiagnosticsProperty('style', style))
-      ..add(DiagnosticsProperty('selected', selected))
-      ..add(DiagnosticsProperty('date', date))
-      ..add(DiagnosticsProperty('today', today));
-  }
-}
 
 /// [FLineCalendar]'s style.
 final class FLineCalendarStyle with Diagnosticable {
