+++ forui/lib/widgets/line_calendar.dart
@@ -1,0 +1,10 @@
+/// {@category Widgets}
+///
+/// A line calendar that allows users to select a single date.
+///
+/// See https://forui.dev/docs/line-calendar for working examples.
+library forui.widgets.lineCalendar;
+
+export '../src/widgets/line_calendar/line_calendar.dart';
+export '../src/widgets/line_calendar/line_calendar_controller.dart';
+export '../src/widgets/line_calendar/line_calendar_tile.dart';
