+++ forui/test/src/widgets/line_calendar_golden_test.dart
@@ -5,7 +5,7 @@ import 'package:flutter/widgets.dart';
 
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:forui/src/widgets/line_calendar.dart';
+import 'package:forui/src/widgets/line_calendar/line_calendar.dart';
 import '../test_scaffold.dart';
 
 void main() {
